actions authentication

+++ src/lib/actions.ts
@@ -4,9 +4,13 @@ import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
 import { songSchema } from "@/lib/schemas";
 import { Song } from "@/lib/schemas";
-import { signIn, signOut } from "@/auth";
+import { signIn, signOut, auth } from "@/auth";
 
 export async function createSong(song: Song) {
+  const session = await auth();
+  if (!session) {
+    throw new Error("You must be signed in to perform this action");
+  }
   const validatedFields = songSchema.safeParse(song);
   if (!validatedFields.success) {
     return {
@@ -32,8 +36,11 @@ export async function createSong(song: Song) {
 }
 
 export async function updateSong(song: Song, id: string) {
+  const session = await auth();
+  if (!session) {
+    throw new Error("You must be signed in to perform this action");
+  }
   const { measures } = song;
-
   try {
     await sql`
       INSERT INTO songs (measures)
@@ -51,6 +58,10 @@ export async function updateSong(song: Song, id: string) {
 }
 
 export async function deleteSong(id: string) {
+  const session = await auth();
+  if (!session) {
+    throw new Error("You must be signed in to perform this action");
+  }
   try {
     await sql`DELETE FROM songs WHERE id = ${id}`;
   } catch (error) {
